Move service code to general class
