+++ flutter-idea/src/io/flutter/devtools/DevToolsViewService.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2024 The Chromium Authors. All rights reserved.
+ * Use of this source code is governed by a BSD-style license that can be
+ * found in the LICENSE file.
+ */
+package io.flutter.devtools;
+
+import com.intellij.openapi.project.Project;
+import io.flutter.view.EmbeddedBrowser;
+import org.jetbrains.annotations.NotNull;
+
+/**
+ * Common functionality for a DevTools view embedded in a tool window.
+ */
+public abstract class DevToolsViewService {
+  private final Project myProject;
+  private EmbeddedBrowser embeddedBrowser;
+
+  DevToolsViewService(Project project) {
+    this.myProject = project;
+  }
+
+  public void setEmbeddedBrowser(EmbeddedBrowser embeddedBrowser) {
+    this.embeddedBrowser = embeddedBrowser;
+  }
+
+  public void updateVmServiceUri(@NotNull String vmServiceUri) {
+    if (this.embeddedBrowser == null) return;
+    this.embeddedBrowser.updateVmServiceUri(vmServiceUri);
+  }
+}
