+++ flutter-idea/src/io/flutter/devtools/RemainingDevToolsViewService.java
@@ -11,20 +11,8 @@ import io.flutter.view.EmbeddedBrowser;
 import org.jetbrains.annotations.NotNull;
 
 @Service(Service.Level.PROJECT)
-public final class RemainingDevToolsViewService {
-  private final Project myProject;
-  private EmbeddedBrowser embeddedBrowser;
-
-  RemainingDevToolsViewService(Project project) {
-    this.myProject = project;
-  }
-
-  public void setEmbeddedBrowser(EmbeddedBrowser embeddedBrowser) {
-    this.embeddedBrowser = embeddedBrowser;
-  }
-
-  public void updateVmServiceUri(@NotNull String vmServiceUri) {
-    if (this.embeddedBrowser == null) return;
-    this.embeddedBrowser.updateVmServiceUri(vmServiceUri);
+public final class RemainingDevToolsViewService extends DevToolsViewService {
+  RemainingDevToolsViewService(@NotNull Project project) {
+    super(project);
   }
 }
